+++ tests/testthat/test-all.R
@@ -22,6 +22,15 @@ test_that("get_eurostat return right classes",{
                            class)))
 })
 
+test_that("get_eurostat handles daily data", {
+  skip_on_cran()
+  dat <- get_eurostat("ert_bil_eur_d", time_format = "date", cache = FALSE)
+  dat1 <- subset(dat, currency == "ARS")
+  expect_equal(as.numeric(difftime(dat1$time[1], dat1$time[2], units = "days")), 1)
+})
+
+
+
 context("cache")
 
 test_that("Cache give error if cache dir does not exist", {